fix: Resolve 20+ TODO/FIXME items — compiler warnings, dead code, typos, and code quality (v9.54)
Compiler warning fixes (7):
- Remove dead SSCMMode enum + getSSCMMode() from client.cpp (-Wunused-function)
- Fix sign-compare in test_crypto.cpp: UASSERTEQ(int,...) -> UASSERTEQ(u8,...) (2 instances)
- Fix unused-but-set-variable in test_encrypted_packet_format.cpp: restructure decrypt logic
- Fix unused-but-set-variable in test_peer_encryption_state.cpp: add (void) casts (3 instances)

Dead code / cleanup fixes (5):
- Remove unused Meta struct from profilergraph.h (dead code)
- Fix duplicate #include gui/guiPasswordChange.h in game_formspec.cpp
- Fix 'fromspec' -> 'formspec' typos in game_formspec.h (3 instances)
- Fix 'intend' -> 'is intended' in game_formspec.h
- Fix 'interations' -> 'iterations' typos in collision.cpp (3 instances)

FIXME/TODO resolution fixes (9):
- Add palette cache rebuild to TextureSource::rebuildImages() (fixes FIXME)
- Add type validation to read_ARGB8() mirroring is_color_table() (fixes FIXME)
- Resolve test_collision.cpp FIXME: document ignore-block behavior as expected
- Resolve test_sao.cpp FIXME: document deactivation step as expected behavior
- Improve game_formspec.cpp Android FIXME: document m_formspec sync limitation
- Add DEPRECATED annotation to m_formspec in game_formspec.h (resolves TODO)
- Make sound speed a named constexpr in sound_singleton.cpp (fixes FIXME)
- Add ownership docs + null assertion to DecoSchematic::clone() (fixes FIXME)
- Improve renderingengine.h TODO with migration guidance

+++ src/unittest/test_crypto.cpp
@@ -823,7 +823,7 @@ void TestCrypto::testBuildNonceCounterZero()
 
         // Last 8 bytes should be all zeros (counter = 0, big-endian)
         for (size_t i = NONCE_BASE_SIZE; i < GCM_NONCE_SIZE; i++) {
-                UASSERTEQ(int, nonce[i], 0);
+                UASSERTEQ(u8, nonce[i], u8(0));
         }
 }
 
@@ -840,7 +840,7 @@ void TestCrypto::testBuildNonceCounterMax()
 
         // Last 8 bytes should be all 0xFF (counter = max, big-endian)
         for (size_t i = NONCE_BASE_SIZE; i < GCM_NONCE_SIZE; i++) {
-                UASSERTEQ(int, nonce[i], 0xFF);
+                UASSERTEQ(u8, nonce[i], u8(0xFF));
         }
 }
